# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agentic-octopus",
    "description": "\nThe **spine** for personal agentic applications: one orchestration service every agent\nplugs into \u2014 triggering (API + cron schedules), a durable Postgres-backed run queue,\nhuman **approval gates**, full audit trail, and multi-surface **chat**.\n\n### Authentication\nAll routes except `/healthz` require `Authorization: Bearer <token>`:\n\n| Token | Scope |\n|---|---|\n| `OCTO_API_TOKEN` | admin \u2014 everything |\n| `OCTO_CHAT_TOKEN` | chat surfaces only (`/chat/*`, `/v1/*`) \u2014 hand this to chat UIs |\n\n### Model routing\n`octo/auto` (the default) smart-routes across healthy **free** models with server-side\nfallback. Explicit model ids are honored as-is; non-free models are refused unless the\noperator opted in (`OPENROUTER_ALLOW_PAID`) or selected `octo/claude` with an Anthropic\nkey configured.\n\n### Surfaces\n- **Native API** (this spec): agents, runs, approvals, schedules, chat.\n- **`/v1`** \u2014 OpenAI-compatible *protocol namespace* (not a version): plug in any\n  OpenAI-protocol client (Open WebUI, SDKs). Plain completions; tool fields stripped.\n\nFull design: `docs/rfcs/0001-agentic-octopus-the-spine.md` \u00b7 Decisions: `docs/adr/`\n",
    "contact": {
      "name": "Diego Silva",
      "url": "https://gitlab.com/behold-corp/agentic-octopus"
    },
    "license": {
      "name": "Private"
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/chat/conversations": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Conversation",
        "operationId": "create_conversation_chat_conversations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Conversations",
        "operationId": "list_conversations_chat_conversations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Conversation",
        "operationId": "get_conversation_chat_conversations__conversation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Delete Conversation",
        "operationId": "delete_conversation_chat_conversations__conversation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat/conversations/{conversation_id}/messages": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Messages",
        "operationId": "get_messages_chat_conversations__conversation_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Message",
        "operationId": "send_message_chat_conversations__conversation_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat/usage": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Usage",
        "operationId": "usage_chat_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/models": {
      "get": {
        "tags": [
          "openai-compat"
        ],
        "summary": "List Models",
        "operationId": "list_models_v1_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "openai-compat"
        ],
        "summary": "Chat Completions",
        "operationId": "chat_completions_v1_chat_completions_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConversationRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "default": "default"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          }
        },
        "type": "object",
        "title": "ConversationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MessageRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "OCTO_CHAT_TOKEN (chat scope) or OCTO_API_TOKEN (admin)",
        "scheme": "bearer"
      }
    }
  },
...
}